--- scripts/cell_fluxes.csv
@@ -1,17 +1,0 @@
-x_center,y_center,center_flux
-0.5,0.5,2.0
-0.5,1.5,1.5
-0.5,2.5,1.0
-0.5,3.5,0.5
-1.5,0.5,1.75
-1.5,1.5,1.25
-1.5,2.5,0.75
-1.5,3.5,0.25
-2.5,0.5,1.5
-2.5,1.5,1.0
-2.5,2.5,0.5
-2.5,3.5,0.0
-3.5,0.5,1.25
-3.5,1.5,0.75
-3.5,2.5,0.25
-3.5,3.5,-0.25
